# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/v1/auth/otp/request": {
      "post": {
        "description": "Génère un code à 6 chiffres, le stocke haché dans `otp_codes` (expiration 5 minutes, 5 tentatives) et l'envoie par SMS. La réponse est identique que le numéro soit connu ou non. Limitation : 3 demandes / 10 min par numéro, 20 / heure par adresse IP.",
        "operationId": "AuthController_requestOtp",
        "parameters": [
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "IAM.PHONE_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "IAM.RATE_LIMITED ou IAM.OTP_RESEND_TOO_SOON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Demander un code de connexion",
        "tags": ["Authentification"]
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "description": "Crée le compte au premier succès. Émet un JWT d'accès (15 min) et un refresh token opaque (30 jours). À la 5e erreur, le code est verrouillé, une entrée `OTP_LOCKED` est écrite dans `audit_logs` et l'API répond 429.",
        "operationId": "AuthController_verifyOtp",
        "parameters": [
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "IAM.OTP_INVALID / IAM.OTP_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "IAM.OTP_LOCKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Vérifier le code et ouvrir une session",
        "tags": ["Authentification"]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Rotation systématique : le jeton présenté est révoqué et un nouveau est émis dans la même famille. Le rejeu d'un jeton déjà révoqué révoque TOUTE la famille et répond 401 `IAM.REFRESH_REVOKED`.",
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "IAM.REFRESH_INVALID / IAM.REFRESH_EXPIRED / IAM.REFRESH_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Faire tourner le refresh token",
        "tags": ["Authentification"]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Révoque toute la famille de jetons de l'appareil. Idempotent.",
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session fermée."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Se déconnecter",
        "tags": ["Authentification"]
      }
    },
...
    "/v1/organizations/{id}/invitations/{invitationId}": {
      "delete": {
        "operationId": "OrganizationsController_revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annuler une invitation",
        "tags": ["Organisations"]
      }
    },
    "/v1/invitations/{token}": {
      "get": {
        "description": "Route publique : la connaissance du jeton (32 octets) vaut autorisation de lire l'entête de l'invitation. Aucune donnée métier de l'organisation n'est exposée.",
        "operationId": "InvitationsController_peek",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicInvitationDto"
                }
              }
            }
          },
          "404": {
            "description": "INVITATION.NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "410": {
            "description": "INVITATION.EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Consulter une invitation avant de l'accepter",
        "tags": ["Invitations"]
      }
    },
    "/v1/invitations/{token}/accept": {
      "post": {
        "description": "L'utilisateur connecté doit être le titulaire du numéro invité. L'acceptation est atomique : un lien ne peut être consommé qu'une fois.",
        "operationId": "InvitationsController_accept",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedMembershipDto"
                }
              }
            }
          },
          "403": {
            "description": "INVITATION.PHONE_MISMATCH",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "INVITATION.ALREADY_USED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accepter une invitation",
        "tags": ["Invitations"]
      }
    },
...
  },
  "info": {
    "title": "Immodesk API",
    "description": "API de la plateforme Immodesk (gestion immobilière, Congo-Brazzaville).\n\n**Conventions**\n- Préfixe : `/v1`. Dates ISO 8601 UTC, montants entiers XAF (BigInt sérialisé en chaîne).\n- Authentification : `Authorization: Bearer <accessToken>` (JWT, 15 minutes).\n- Contexte d'organisation : en-tête `X-Organization-Id` obligatoire sur toute route d'organisation.\n- Erreurs : `{ code, message, details }`, codes stables `DOMAINE.RAISON`, messages en français (fr-CG).\n- Toute ressource appartenant à une autre organisation répond **404**, jamais 403.\n- Pagination par curseur : `?limit=50&cursor=...` → `{ items, pageInfo }`.\n- Idempotence : en-tête facultatif `Idempotency-Key` sur les POST.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Jeton d'accès (15 min)."
      }
    },
    "schemas": {
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+242066000001",
            "description": "Numéro de téléphone. Les formes nationales (066000001, 06 600 00 01) et internationales (00242..., 242...) sont normalisées en E.164 avec l'indicatif +242."
          },
          "channel": {
            "type": "string",
            "enum": ["SMS", "WHATSAPP"],
            "default": "SMS"
          }
        },
        "required": ["phone"]
      },
      "OtpRequestResponseDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "channel": {
            "type": "string",
            "enum": ["SMS", "WHATSAPP"]
          },
          "expiresInSeconds": {
            "type": "number",
            "example": 300
          },
          "resendAfterSeconds": {
            "type": "number",
            "example": 60
          }
        },
        "required": ["requestId", "channel", "expiresInSeconds", "resendAfterSeconds"]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "IAM.OTP_INVALID",
            "description": "Code stable `DOMAINE.RAISON`."
          },
          "message": {
            "type": "string",
            "example": "Code incorrect.",
            "description": "Message en français (fr-CG)."
          },
          "details": {
            "type": "object",
            "example": {
              "remainingAttempts": 3
            }
          }
        },
        "required": ["code", "message", "details"]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+242066000001"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Code à 6 chiffres reçu par SMS."
          },
          "deviceName": {
            "type": "string",
            "example": "Samsung A14 — Brazzaville"
          }
        },
        "required": ["phone", "code"]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "string",
            "example": "+242066000001"
          },
          "fullName": {
            "type": "string",
            "example": "Jean Mabiala"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "example": "fr-CG"
          },
          "timezone": {
            "type": "string",
            "example": "Africa/Brazzaville"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "phone", "fullName", "email", "locale", "timezone", "createdAt"]
      },
      "OrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["AGENCY", "INDEPENDENT_LANDLORD", "INDEPENDENT_MANAGER"]
          },
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "example": "Brazzaville"
          },
          "district": {
            "type": "string",
            "nullable": true
          },
          "contactPhone": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "legalName",
          "tradeName",
          "slug",
          "city",
          "district",
          "contactPhone",
          "contactEmail",
          "logoUrl",
          "status",
          "createdAt"
        ]
      },
      "OrganizationMembershipDto": {
        "type": "object",
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationDto"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "COLLECTOR", "ACCOUNTANT", "VIEWER"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["organization", "role", "joinedAt"]
      },
      "OtpVerifyResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationMembershipDto"
            }
          }
        },
        "required": ["accessToken", "refreshToken", "user", "organizations"]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token opaque émis par /v1/auth/otp/verify."
          }
        },
        "required": ["refreshToken"]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": ["accessToken", "refreshToken"]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token de l'appareil à déconnecter."
          }
        },
        "required": ["refreshToken"]
      },
...
      "PublicInvitationDto": {
        "type": "object",
        "properties": {
          "organizationName": {
            "type": "string",
            "example": "Agence Mpila Immo"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "ACCOUNTANT", "COLLECTOR", "VIEWER"]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["organizationName", "role", "expiresAt"]
      },
      "AcceptedMembershipDto": {
        "type": "object",
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationDto"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "ACCOUNTANT", "COLLECTOR", "VIEWER"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["organization", "role", "joinedAt"]
      },
...
    }
  }
}